+ Swing Reserve: [fixes #16] - fixes Player::m_swing_dice[] not being updated when a swing reserve die is used (i.e. goes from !IsUsed() to IsUsed()) - EXPL: there is a co-dependency between the m_swing_dice[] counts and the Die::IsUsed() state. See BMC_Player::OnDiceParsed(). If m_swing_dice[] is not updated, then the AI doesn't know that it needs to include that swing type in it's list of candidate SetSwing actions. - added test input for bug + Warnings: fixed one size_t -> INT cast warning + Style: fixed several inconsistent if/else bracketing + Project: - partial work towards splitting up classes into distinct headers/source files - headers: setup several - source: only setup Stats.cpp - player.cpp: renamed to drop the prefix
