+++ src/bmai.h
@@ -7,8 +7,11 @@
 // DESC: main header for BMAI
 //
 // REVISION HISTORY:
+// drp030321 - partial split out to individual headers
+//
+// TODO:
+// 1) drp030321 - setup a main precompiled header that includes everything (bmai.h) vs a header for the key types/enums/classes. Split out modules
 ///////////////////////////////////////////////////////////////////////////////////////////
-
 
 #ifndef __BMAI_H__
 #define __BMAI_H__
@@ -272,6 +275,12 @@ extern class BMC_Logger	g_logger;
 extern class BMC_QAI	g_qai;
 extern class BMC_Stats	g_stats;
 
+// MOOD dice
+#define BMD_MOOD_SIDES_RANGE_X	6
+#define BMD_MOOD_SIDES_RANGE_V	4
+extern INT g_mood_sides_X[BMD_MOOD_SIDES_RANGE_X];
+extern INT g_mood_sides_V[BMD_MOOD_SIDES_RANGE_V];
+
 ///////////////////////////////////////////////////////////////////////////////////////////
 // classes
 ///////////////////////////////////////////////////////////////////////////////////////////
@@ -388,399 +397,11 @@ private:
 };
 */
 
-// NOTES:
-// - second die only used for Option and Twin
-class BMC_DieData
-{
-	friend class BMC_Parser;
 
-public:
-				BMC_DieData();
-	virtual void Reset();
 
-	// methods
-	void		Debug(BME_DEBUG _cat = BME_DEBUG_ALWAYS);
 
-	// accessors
-	bool		HasProperty(INT _p)  { return m_properties & _p; }
-	BME_SWING	GetSwingType(INT _d) { return (BME_SWING)m_swing_type[_d]; }
-	bool		Valid()				 { return (m_properties & BME_PROPERTY_VALID); }
-	// drp022521 - fixed to return INT instead of bool
-	INT			Dice()				 { return (m_properties & BME_PROPERTY_TWIN) ? 2 : 1; }
-	INT			GetSides(INT _t)	{ return m_sides[_t]; }
 
-protected:
-	U32			m_properties;
-	U8			m_sides[BMD_MAX_TWINS];			// number of sides if not a swing die, or current number of sides
 
-private:
-	U8			m_swing_type[BMD_MAX_TWINS];	// definition number of sides, should be BME_SWING
-};
-
-class BMC_Die : public BMC_DieData
-{
-	friend		class BMC_Parser;
-
-public:
-	// setup
-	virtual void Reset();
-	void		SetDie(BMC_DieData *_data);
-	void		Roll();
-	void		GameRoll(BMC_Player *_owner);
-
-	// methods
-	void		Debug(BME_DEBUG _cat = BME_DEBUG_ALWAYS);
-	void		SetOption(INT _d);
-	void		SetFocus(INT _v);
-
-	// accessors
-	bool		CanDoAttack(BMC_MoveAttack &_move)	{ return m_attacks.IsSet(_move.m_attack); }
-	bool		CanBeAttacked(BMC_MoveAttack &_move)	{ return m_vulnerabilities.IsSet(_move.m_attack); }
-	INT			GetValueTotal()					{ return m_value_total; }
-	INT			GetSidesMax()					{ return m_sides_max; }
-	bool		IsAvailable()					{ return m_state == BME_STATE_READY || m_state == BME_STATE_DIZZY; }
-	bool		IsInReserve()					{ return m_state == BME_STATE_RESERVE; }
-	bool		IsUsed()						{ return m_state != BME_STATE_NOTUSED && m_state!=BME_STATE_RESERVE; }
-	float		GetScore(bool _own);
-	INT			GetOriginalIndex()				{ return m_original_index; }
-	BME_STATE	GetState()						{ return (BME_STATE)m_state; }
-
-	// mutators
-	void		SetState(BME_STATE _state)		{ m_state = _state; }
-	void		CheatSetValueTotal(INT _v)		{ m_value_total = _v; }	// used for some functions
-
-	// events
-	void		OnDieChanged();
-	void		OnSwingSet(INT _swing, U8 _value);
-	void		OnApplyAttackPlayer(BMC_Move &_move, BMC_Player *_owner, bool _actually_attacking=true);
-	void		OnApplyAttackNatureRollAttacker(BMC_Move &_move, BMC_Player *_owner);
-	void		OnApplyAttackNatureRollTripped();
-	void		OnBeforeRollInGame(BMC_Player *_owner);
-	void		OnUseReserve();
-	void		OnDizzyRecovered();
-
-protected:
-
-	// call this once state changes
-	void		RecomputeAttacks();
-
-private:
-	U8			m_state;				// should be BME_STATE
-	U8			m_value_total;			// current total value of all dice, redundant
-	U8			m_sides_max;			// max value of dice (based on m_sides), redundant
-	U8			m_original_index;		// save original index for interface
-	BMC_BitArray<BME_ATTACK_MAX>	m_attacks;
-	BMC_BitArray<BME_ATTACK_MAX>	m_vulnerabilities;
-	//U8			m_value[BMD_MAX_TWINS];	// current value of dice, not really necessary (and often not known!)
-};
-
-class BMC_Man
-{
-public:
-	// accessors
-	BMC_DieData *GetDieData(INT _d) { return &m_die[_d]; }
-
-protected:
-private:
-	BMC_DieData	m_die[BMD_MAX_DICE];
-};
-
-class BMC_Player	// 204b
-{
-	friend class BMC_Parser;
-
-public:
-	typedef enum {
-		SWING_SET_NOT,
-		SWING_SET_READY,	// set this round, esults should not be "known" to opponent - to simulate simultaneous swing set
-		SWING_SET_LOCKED,	// set from previous round
-	} SWING_SET;
-
-				BMC_Player();
-	void		SetID(INT _id) { m_id = _id; }			
-	void		Reset();				
-	void		OnDiceParsed();
-
-	// playing with normal rules
-	void		SetButtonMan(BMC_Man *_man);
-	void		SetSwingDice(INT _swing, U8 _value, bool _from_turbo=false);
-	void		SetOptionDie(INT _i, INT _d);
-	void		RollDice();
-
-	// methods
-	void		Debug(BME_DEBUG _cat = BME_DEBUG_ALWAYS);
-	void		DebugAllDice(BME_DEBUG _cat = BME_DEBUG_ALWAYS);
-	void		SetSwingDiceStatus(SWING_SET _swing)	{ m_swing_set = _swing; }
-	bool		NeedsSetSwing();
-
-	// events
-	void		OnDieSidesChanging(BMC_Die *_die);
-	void		OnDieSidesChanged(BMC_Die *_die);
-	BMC_Die *	OnDieLost(INT _d);
-	void		OnDieCaptured(BMC_Die *_die);
-	void		OnRoundLost();
-	void		OnSurrendered();
-	//void		OnSwingDiceSet() { m_swing_set = true; }
-	void		OnSwingDiceReady()	{ m_swing_set = SWING_SET_READY; }
-	void		OnAttackFinished() { OptimizeDice(); }
-	void		OnDieTripped() { OptimizeDice(); }
-	void		OnChanceDieRolled() { OptimizeDice(); }
-	void		OnFocusDieUsed()	{ OptimizeDice(); }
-
-	// accessors
-	BMC_Die *	GetDie(INT _d) { return &m_die[_d]; }
-	INT			GetAvailableDice() { return m_available_dice; }
-	INT			GetMaxValue() { return m_max_value; }
-	float		GetScore() { return m_score; }
-	//bool		SwingDiceSet() { return m_swing_set; }
-	SWING_SET	GetSwingDiceSet()	{ return m_swing_set; }
-	INT			HasDieWithProperty(INT _p, bool _check_all_dice=false);
-	INT			GetTotalSwingDice(INT _s) { return m_swing_dice[_s]; }
-	INT			GetID() { return m_id; }
-
-
-protected:
-	// methods
-	void		OptimizeDice();
-
-private:
-	BMC_Man	*	m_man;
-	INT			m_id;
-	SWING_SET	m_swing_set;
-	BMC_Die		m_die[BMD_MAX_DICE];			// as long as Optimize was called, these are sorted largest to smallest that are READY
-	U8			m_swing_value[BME_SWING_MAX];	
-	U8			m_swing_dice[BME_SWING_MAX];	// number of dice of each swing type
-	INT			m_available_dice;				// only valid after Optimize
-	INT			m_max_value;					// only valid after Optimize, useful to know for skill attacks
-	float		m_score;
-};
-
-class BMC_Game		// 420b
-{
-	friend class BMC_Parser;
-
-public:
-				BMC_Game(bool _simulation);
-				BMC_Game(const BMC_Game & _game);
-	const BMC_Game & operator=(const BMC_Game &_game);
-
-	// game simulation - level 0
-	void		PlayGame(BMC_Man *_man1 = NULL, BMC_Man *_man2 = NULL);
-
-	// game simulation - level 1 (for simulations)
-	BME_WLT		PlayRound(BMC_Move *_start_action=NULL);
-
-	// game methods
-	bool		ValidAttack(BMC_MoveAttack &_move);
-	void		GenerateValidAttacks(BMC_MoveList &_movelist);
-
-	bool		ValidSetSwing(BMC_Move &_move);
-	void		GenerateValidSetSwing(BMC_MoveList & _movelist);
-	void		ApplySetSwing(BMC_Move &_move, bool _lock = true);
-
-	bool		ValidUseFocus(BMC_Move &_move);
-	void		GenerateValidFocus(BMC_MoveList & _movelist);
-	void		ApplyUseFocus(BMC_Move &_move);
-
-	void		ApplyUseReserve(BMC_Move &_move);
-
-	bool		ValidUseChance(BMC_Move &_move);
-	void		GenerateValidChance(BMC_MoveList & _movelist);
-	void		ApplyUseChance(BMC_Move &_move);
-
-	// initiative
-	INT			CheckInitiative();
-
-	// managing fights
-	bool		FightOver();
-	void		ApplyAttackPlayer(BMC_Move &_move);
-	void		ApplyAttackNatureRoll(BMC_Move &_move);
-	void		ApplyAttackNaturePost(BMC_Move &_move, bool &_extra_turn);
-	void		SimulateAttack(BMC_MoveAttack &_move, bool & _extra_turn);
-	void		RecoverDizzyDice(INT _player);
-
-	// accessors
-	BMC_Player *GetPlayer(INT _i) { return &m_player[_i]; }
-	BMC_Player *GetPhasePlayer() { return GetPlayer(m_phase_player); }
-	BMC_Player *GetTargetPlayer() { return GetPlayer(m_target_player); }
-	INT			GetPhasePlayerID() { return m_phase_player; }
-	bool		IsPreround() { return m_phase == BME_PHASE_PREROUND || m_phase==BME_PHASE_RESERVE; }
-	BME_PHASE	GetPhase() { return m_phase; }
-	INT			GetStanding(INT _wlt) { return m_standing[_wlt]; }
-	INT			GetInitiativeWinner() { return m_initiative_winner; }
-	bool		IsSimulation() { return m_simulation; }
-	BMC_AI *	GetAI(INT _p) { return m_ai[_p]; }
-
-	// mutators
-	void		SetAI(INT _p, BMC_AI *_ai) { m_ai[_p] = _ai; }
-
-	// methods wrt. "percent chance to win"
-	float		ConvertWLTToWinProbability();
-	float		PlayFight_EvaluateMove(INT _pov_player, BMC_Move &_move);
-	float		PlayRound_EvaluateMove(INT _pov_player);
-
-protected:
-	// game simulation - level 1
-	void		Setup(BMC_Man *_man1 = NULL, BMC_Man *_man2 = NULL);
-
-	// game simulation - level 2
-	void		PlayPreround();
-	void		PlayInitiative();
-	void		PlayInitiativeChance();
-	void		PlayInitiativeFocus();
-	void		PlayFight(BMC_Move *_start_action=NULL);
-	void		FinishPreround();
-	void		FinishInitiative();
-	void		FinishInitiativeChance(bool _swap_phase_player);
-	void		FinishInitiativeFocus(bool _swap_phase_player);
-	BME_WLT		FinishRound(BME_WLT _wlt_0);
-
-	// game simulation - level 3
-	void		FinishTurn(bool extra_turn=false);
-
-private:
-	BMC_Player	m_player[BMD_MAX_PLAYERS];
-	U8			m_standing[BME_WLT_MAX];
-	U8			m_target_wins;
-	BME_PHASE	m_phase;
-	U8			m_phase_player;
-	U8			m_target_player;
-	BME_ACTION	m_last_action;
-
-	// AI players
-	BMC_AI *	m_ai[BMD_MAX_PLAYERS];
-
-	// stats accumulated during the round
-	U8			m_initiative_winner;	// the player who originally won initiative (ignoring CHANCE and FOCUS)
-
-	// is this a simulation run by the AI?
-	bool		m_simulation;
-};
-
-
-// INVARIANT: only contains dice in increasing order (e.g. 0,2,3 but not 2.0,3)				
-class BMC_DieIndexStack
-{
-public:
-	BMC_DieIndexStack(BMC_Player *_owner);
-
-	// mutators
-	void			Pop()	{ die_stack_size--; }
-	void			Push(INT _index);
-	bool			Cycle(bool _add_die=true);
-
-	// methods 
-	void			SetBits(BMC_BitArray<BMD_MAX_DICE> & _bits);
-	void			Debug(BME_DEBUG _cat = BME_DEBUG_ALWAYS);
-
-	// accessors
-	INT				GetValueTotal() { return value_total; }
-	bool			ContainsAllDice() { return die_stack_size == owner->GetAvailableDice(); }
-	bool			ContainsLastDie() { return GetTopDieIndex() == owner->GetAvailableDice()-1; }
-	INT				GetTopDieIndex() { BM_ASSERT(die_stack_size>0); return die_stack[die_stack_size-1]; }
-	BMC_Die *		GetTopDie()		{ return owner->GetDie(GetTopDieIndex()); }
-	bool			Empty()			{ return die_stack_size == 0; }
-	INT				GetStackSize()	{ return die_stack_size; }
-	BMC_Die *		GetDie(int _index) { return owner->GetDie(die_stack[_index]); }
-	INT				CountDiceWithProperty(BME_PROPERTY _property);
-
-protected:
-private:
-	BMC_Player *	owner;
-	INT				die_stack[BMD_MAX_DICE];
-	INT				die_stack_size;
-	INT				value_total;
-};
-
-class BMC_Parser
-{
-public:
-	BMC_Parser();
-	void	SetupTestGame();
-	void	ParseGame();
-	void	Parse();
-	void	Parse(FILE *_fp)	{ file = _fp; Parse(); }
-
-protected:
-	void			GetAction();
-	void			PlayGame(INT _games);
-	void			CompareAI(INT _games);
-	void			PlayFairGames(INT _games, INT _mode, F32 _p);
-	void			ParseDie(INT _p, INT _dice);
-	void			ParsePlayer(INT _p, INT _dice);
-	bool			Read(bool _fatal=true);
-
-	// output
-	void			Send(char *_fmt, ...);
-	void			SendStats();
-	void			SendSetSwing(BMC_Move &_move);
-	void			SendUseReserve(BMC_Move &_move);
-	void			SendAttack(BMC_Move &_move);
-	void			SendUseChance(BMC_Move &_move);
-	void			SendUseFocus(BMC_Move &_move);
-
-	// parsing dice
-	bool			DieIsSwing(char _c) { return _c>=BMD_FIRST_SWING_CHAR && _c<=BMD_LAST_SWING_CHAR; }
-	bool			DieIsNumeric(char _c) { return (_c>='0' && _c<='9'); }
-	bool			DieIsValue(char _c) { return DieIsSwing(_c) || DieIsNumeric(_c); }
-	bool			DieIsTwin(char _c) { return _c=='('; }
-	bool			DieIsOption(char _c) { return _c=='/'; }
-
-private:
-	// parsing dice methods (uses 'line')
-	INT				ParseDieNumber(INT & _pos);
-	void			ParseDieSides(INT & _pos, INT _die);
-	INT				ParseDieDefinedSides(INT _pos);
-
-	// state for dice parsing
-	BMC_Player *p;
-	BMC_Die *d;
-	char			line[BMD_MAX_STRING];
-	FILE *file;
-};
-
-class BMC_Logger
-{
-public:
-					BMC_Logger();
-	
-	// methods
-	void			Log(BME_DEBUG _cat, char *_fmt, ... );
-
-	// mutators
-	void			SetLogging(BME_DEBUG _cat, bool _log) { m_logging[_cat] = _log; }
-	bool			SetLogging(const char *_catname, bool _log);
-
-	// accessors
-	bool			IsLogging(BME_DEBUG _cat) { return m_logging[_cat]; }
-	
-private:
-	bool			m_logging[BME_DEBUG_MAX];
-};
-
-class BMC_Stats
-{
-public:
-	BMC_Stats();
-
-	// methods
-	void			DisplayStats();
-
-	// events
-	void			OnAppStarted()	{ m_start = time(NULL); }
-	void			OnFullSimulation() { m_sims++; }
-
-	// bmai-specific
-	void			OnPlyAction(int _ply, int _moves, int _sims) { m_total_sims[_ply]+=_sims; m_total_moves[_ply]+=_moves; m_total_samples[_ply]++; }
-
-private:
-	time_t			m_start, m_end;
-	int				m_sims;
-	int				m_total_sims[BMD_MAX_PLY];
-	int				m_total_moves[BMD_MAX_PLY];
-	int				m_total_samples[BMD_MAX_PLY];
-
-};
 
 // Utility functions
